requirements.txt
asgiref==3.8.1
Brotli==1.1.0
click==8.1.8
colorama==0.4.6
dj-database-url==2.3.0
Django==5.1.7
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
gunicorn==23.0.0
h11==0.14.0
packaging==24.2
pillow==11.1.0
promise==2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
whitenoise==6.9.0
djangorestframework-simplejwt
graphene-file-upload
django-cors-headers